// src/enums/Events.ts
export enum Events {
    Null = 0 << 0,
    Sphinx = 1 << 0,
    LOTLL = 1 << 1,
    Skull = 1 << 2,
    Cube = 1 << 3,
    Avatar = 1 << 4,
    Dragon = 1 << 5,
    Statues = 1 << 6,
    Sentry = 1 << 7,
    GShip = 1 << 8,
    Hive = 1 << 9,
    Hermit = 1 << 10,
    Pentaract = 1 << 11
}

export enum Bags {
    Soulbound = 0x0503,
    Blue = 0x050B, //5
    BlueBoosted = 0x6be,
    Brown = 0x0500, //0
    BrownBoosted = 0x6ad,
    Cyan = 0x0509, //4
    CyanBoosted = 0x6bd,
    Egg = 0x0508, //3
    EggBoosted = 0x6bb,
    Gold = 0x050E, //7
    GoldBoosted = 0x6bc,
    Orange = 0x50F, //8
    OrangeBoosted = 0x6bf,
    Pink = 0x0506, //1
    PinkBoosted = 0x6ae,
    Purple = 0x0507, //2
    PurpleBoosted = 0x6ba,
    Red = 0x6AC, //9
    RedBoosted = 0x6c0,
    White = 0x050C, //6
    WhiteBoosted = 0x0510,
}